reorg: reorg backend api folder/files + dependencies +services/api => main
